how is i16 faster than u8?

+++ santorini_core/src/gods/generic.rs
@@ -10,7 +10,7 @@ pub const RETURN_FIRST_MATE: MoveGenFlags = STOP_ON_MATE | MATE_ONLY;
 pub const GRID_POSITION_SCORES: [MoveScore; 25] = grid_position_builder(0, 1, 2, 3, 4, 5);
 pub const WORKER_HEIGHT_SCORES: [MoveScore; 4] = [0, 20, 50, 21];
 
-pub type MoveScore = u8;
+pub type MoveScore = i16;
 pub type MoveData = u32;
 pub const MOVE_IS_WINNING_MASK: MoveData = MoveData::MAX ^ (MoveData::MAX >> 1);
